Std::function
std::function and std::bind were born inside the Boost C++ Library, but they were incorporated into the new C++11 standard.

std::function is a STL template class that provides a very convenient wrapper to a simple function, to a functor or to a lambda expression.


For example, if you want to store several functions, functors or lambda expressions in a vector, you could write something like this:  ^^^



As you can see, in this declaration:


vector<function<void ()>> x;

we are declaring a vector of functions. The void () part means that the functions do not receive any argument and have void as return type. If you would want to define a function that receives two integers and return an integer too, you should declare your function as:

int my_func(int a, int b) { return a + b; } 
function<int (int, int)> f = my_func;